@@ -17,6 +17,7 @@
     "@palxp/color-picker": "^1.5.5",
     "@palxp/image-extraction": "^1.2.4",
     "@scena/guides": "^0.18.1",
+    "@types/cropperjs": "^1.3.0",
     "@webtoon/psd": "^0.4.0",
     "axios": "^0.21.1",
     "core-js": "^3.6.5",
@@ -31,39 +32,30 @@
     "qr-code-styling": "^1.6.0-rc.1",
     "selecto": "^1.13.0",
     "throttle-debounce": "^3.0.1",
-    "vue": "^3.0.0",
+    "vite-plugin-compression": "^0.5.1",
+    "vue": "3.4.19",
     "vue-router": "^4.0.0-0",
     "vuedraggable": "^4.1.0",
     "vuex": "^4.0.0-0"
   },
   "devDependencies": {
-    "@types/node": "^16.3.1",
+    "@types/fontfaceobserver": "^2.1.3",
+    "@types/node": "^20.11.24",
     "@types/throttle-debounce": "^2.1.0",
-    "@typescript-eslint/eslint-plugin": "^4.28.3",
-    "@typescript-eslint/parser": "^4.28.3",
-    "@vitejs/plugin-vue": "^1.2.4",
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-plugin-typescript": "~4.5.0",
-    "@vue/cli-plugin-vuex": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/compiler-sfc": "^3.1.4",
-    "@vue/eslint-config-typescript": "^7.0.0",
+    "@typescript-eslint/eslint-plugin": "^7.1.0",
+    "@typescript-eslint/parser": "^7.1.1",
+    "@vitejs/plugin-vue": "^5.0.4",
     "autoprefixer": "^10.3.1",
-    "babel-eslint": "^10.1.0",
     "cross-env": "^7.0.3",
-    "esbuild-loader": "^2.13.1",
-    "eslint": "^7.29.0",
+    "eslint": "^8.56.0",
     "eslint-config-alloy": "~4.1.0",
     "eslint-plugin-vue": "^7.12.1",
     "less": "^4.1.1",
-    "typescript": "~4.1.5",
+    "terser": "^5.28.1",
+    "typescript": "^5.2.2",
     "unplugin-element-plus": "^0.7.1",
-    "vite": "^2.4.1",
-    "vite-plugin-compression": "^0.3.0",
-    "vue-cli-plugin-norm": "~1.2.2",
-    "vue-eslint-parser": "^7.6.0",
-    "vue-tsc": "^0.2.0"
+    "vite": "^5.1.4",
+    "vue-tsc": "^1.8.27"
   },
   "browserslist": [
     "> 1%",
