@@ -2,9 +2,17 @@
  * @Author: ShawnPhang
  * @Date: 2022-02-12 11:08:57
  * @Description:
- * @LastEditors: ShawnPhang <https://m.palxp.cn>
- * @LastEditTime: 2023-09-19 17:35:44
+ * @LastEditors: ShawnPhang <https://m.palxp.cn>, Jeremy Yu <https://github.com/JeremyYu-cn>
+ * @LastEditTime: 2024-03-01 20:55:51
  */
+
+export type AlignListData = {
+  key: string
+  icon: string
+  tip: string
+  value: string
+}
+
 export default [
   {
     key: 'align',
@@ -42,4 +50,4 @@ export default [
     tip: '下对齐',
     value: 'bottom',
   },
-]
+] as AlignListData[]
