@@ -2,12 +2,19 @@
  * @Author: ShawnPhang
  * @Date: 2022-03-25 13:43:07
  * @Description: 添加滚动监听
- * @LastEditors: ShawnPhang
- * @LastEditTime: 2022-03-25 14:32:19
+ * @LastEditors: ShawnPhang <site: book.palxp.com>, Jeremy Yu <https://github.com/JeremyYu-cn>
+ * @LastEditTime: 2024-03-02 11:50:00
  */
 import store from '@/store'
-export default function(el: Element | string, cb: Function, altLimit: boolean = true) {
+
+type TAddEventCb = (e: Event) => void
+type TAddEventObj = {
+  attachEvent?: HTMLElement["addEventListener"]
+} & HTMLElement
+
+export default function(el: HTMLElement | string, cb: Function, altLimit: boolean = true) {
   const box = typeof el === 'string' ? document.getElementById(el) : el
+  if (!box) return;
   addEvent(box, 'mousewheel', (e: any) => {
     const ev = e || window.event
     const down = ev.wheelDelta ? ev.wheelDelta < 0 : ev.detail > 0
@@ -16,10 +23,7 @@ export default function(el: Element | string, cb: Function, altLimit: boolean = 
     // } else {
     //   console.log('鼠标滚轮向上++++++++++')
     // }
-    if (altLimit && store.getters.dAltDown) {
-      ev.preventDefault()
-      cb(down)
-    } else if (!altLimit) {
+    if ((altLimit && store.getters.dAltDown) || !altLimit) {
       ev.preventDefault()
       cb(down)
     }
@@ -27,7 +31,7 @@ export default function(el: Element | string, cb: Function, altLimit: boolean = 
   })
 }
 
-function addEvent(obj: any, xEvent: string, fn: Function) {
+function addEvent(obj: TAddEventObj, xEvent: keyof HTMLElementEventMap, fn: TAddEventCb) {
   if (obj.attachEvent) {
     obj.attachEvent('on' + xEvent, fn)
   } else {
