@@ -2,17 +2,17 @@
  * @Author: ShawnPhang
  * @Date: 2022-01-31 10:45:53
  * @Description: 用于修改transform字符串
- * @LastEditors: ShawnPhang
- * @LastEditTime: 2022-02-18 16:54:13
+ * @LastEditors: ShawnPhang <site: book.palxp.com>, Jeremy Yu <https://github.com/JeremyYu-cn>
+ * @LastEditTime: 2024-03-02 11:50:00
  */
-export function getTransformAttribute(target: any, attr: string = '') {
+export function getTransformAttribute(target: HTMLElement, attr: string = '') {
   const tf = target.style.transform
   const iof = tf.indexOf(attr)
   const half = tf.substring(iof + attr.length + 1)
   return half.substring(0, half.indexOf(')'))
 }
 
-export function setTransformAttribute(target: any, attr: string, value: string | number = 0) {
+export function setTransformAttribute(target: HTMLElement, attr: string, value: string | number = 0) {
   const tf = target?.style.transform
   if (!tf) {
     return
@@ -24,10 +24,10 @@ export function setTransformAttribute(target: any, attr: string, value: string |
   target.style.transform = FRONT + value + END
 }
 
-export function getMatrix(params: any) {
+export function getMatrix(params: Record<string, any>) {
   const result = []
   for (const key in params) {
-    if (Object.prototype.hasOwnProperty.call(params, key)) {
+    if (Object.hasOwn(params, key)) {
       result.push(params[key])
     }
   }
