@@ -8,17 +8,17 @@
 // TODO: Group类型比较特殊，所以需要全量循环并判断是否为group
 const arr = ['w-text', 'w-image', 'w-svg', 'w-group', 'w-qrcode']
 
-export function getTarget(currentTarget: any) {
-  let collector: any[] = []
-  let groupTarger: any = null
-  let saveTarger: any = null
+export function getTarget(currentTarget: HTMLElement): Promise<HTMLElement | null> {
+  let collector: string[] = []
+  let groupTarger: HTMLElement | null = null
+  let saveTarger: HTMLElement | null = null
   return new Promise((resolve) => {
-    function findTarget(target: any) {
+    function findTarget(target: HTMLElement | null) {
       if (!target || target.id === 'page-design') {
         if (collector.length > 1) {
           resolve(groupTarger)
         } else {
-          resolve(saveTarger || currentTarget)
+          resolve(saveTarger ?? currentTarget)
         }
         return
       }
@@ -37,12 +37,12 @@ export function getTarget(currentTarget: any) {
   })
 }
 
-export function getFinalTarget(currentTarget: any) {
-  let collector: any[] = []
-  let groupTarger: any = null
-  let saveTarger: any = null
+export function getFinalTarget(currentTarget: HTMLElement) {
+  let collector: string[] = []
+  // let groupTarger: HTMLElement | null = null
+  // let saveTarger: HTMLElement | null = null
   return new Promise((resolve) => {
-    function findTarget(target: any) {
+    function findTarget(target: HTMLElement | null) {
       if (!target || target.id === 'page-design') {
         resolve(target)
         return
@@ -51,8 +51,8 @@ export function getFinalTarget(currentTarget: any) {
 
       collector = collector.concat(
         t.filter((x) => {
-          arr.includes(x) && (saveTarger = target)
-          x === 'w-group' && (groupTarger = target)
+          // arr.includes(x) && (saveTarger = target)
+          // x === 'w-group' && (groupTarger = target)
           return arr.includes(x)
         }),
       )
