@@ -1,136 +1,142 @@
 <template>
   <div v-show="showMenuBg" id="menu-bg" class="menu-bg" @click="closeMenu">
     <ul ref="menuList" class="menu-list" :style="styleObj">
-      <li v-for="(item, index) in menuList.list" :key="index" :class="{ 'menu-item': true, 'disable-menu': dCopyElement.length === 0 && item.type === 'paste' }" @click.stop="selectMenu(item.type)">
+      <li v-for="(item, index) in menuListData.list" :key="index" :class="{ 'menu-item': true, 'disable-menu': dCopyElement.length === 0 && item.type === 'paste' }" @click.stop="selectMenu(item.type)">
         {{ item.text }}
       </li>
     </ul>
   </div>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue'
-import { mapGetters, mapActions } from 'vuex'
-import { widgetMenu, pageMenu, menuList } from './rcMenuData'
+<script lang="ts" setup>
+import { computed, onMounted, ref } from 'vue'
+import { useStore } from 'vuex'
+import { 
+  widgetMenu as widget, 
+  pageMenu as page,
+  menuList as menu, 
+  TMenuItemData, TWidgetItemData, 
+} from './rcMenuData'
 import { getTarget } from '@/common/methods/target'
+import { useSetupMapGetters } from '@/common/hooks/mapGetters';
 
-export default defineComponent({
-  setup() {},
-  data() {
-    return {
-      menuList,
-      showMenuBg: false,
-      widgetMenu,
-      pageMenu,
-    }
-  },
-  computed: {
-    ...mapGetters(['dActiveElement', 'dAltDown', 'dWidgets', 'dCopyElement']),
-    styleObj() {
-      return {
-        left: this.menuList.left + 'px',
-        top: this.menuList.top + 'px',
+const store = useStore()
+const menuListData = ref<TMenuItemData>({...menu})
+const showMenuBg = ref<boolean>(false)
+const widgetMenu = ref<TWidgetItemData[]>({...widget})
+const pageMenu = ref<TWidgetItemData[]>({...page})
+
+const {dActiveElement, dAltDown, dWidgets, dCopyElement} = useSetupMapGetters(['dActiveElement', 'dAltDown', 'dWidgets', 'dCopyElement'])
+
+const styleObj = computed(() => {
+  return {
+    left: menuListData.value.left + 'px',
+    top: menuListData.value.top + 'px',
+  }
+})
+
+onMounted(() => {
+  document.oncontextmenu = mouseRightClick
+})
+async function mouseRightClick(e: MouseEvent) {
+  e.stopPropagation()
+  e.preventDefault()
+  if (showMenuBg.value) {
+    showMenuBg.value = false
+    return
+  }
+  if (!e.target) return
+  let target = await getTarget(e.target as HTMLElement)
+  if (!target) return
+  let type = target.getAttribute('data-type')
+  if (type) {
+    let uuid = target.getAttribute('data-uuid') // 设置选中元素
+
+    if (uuid !== '-1' && !dAltDown) {
+      let widget = dWidgets.value.find((item: any) => item.uuid === uuid)
+      if (
+        widget.parent !== '-1' && 
+        widget.parent !== dActiveElement.value.uuid &&
+        widget.parent !== dActiveElement.value.parent
+      ) {
+        uuid = widget.parent
       }
-    },
-  },
-  mounted() {
-    document.oncontextmenu = this.mouseRightClick
-  },
-  methods: {
-    ...mapActions(['selectWidget', 'copyWidget', 'pasteWidget', 'updateLayerIndex', 'deleteWidget', 'ungroup']),
-    async mouseRightClick(e: any) {
-      e.stopPropagation()
-      e.preventDefault()
-      if (this.showMenuBg) {
-        this.showMenuBg = false
+    }
+    store.dispatch('selectWidget', {
+      uuid: uuid ?? '-1',
+    })
+    showMenu(e)
+  }
+}
+
+function showMenu(e: MouseEvent) {
+  let isPage = dActiveElement.value.uuid === '-1'
+  menuListData.value.list = isPage ? pageMenu.value : widgetMenu.value
+  if (dActiveElement.value.isContainer) {
+    let ungroup: TWidgetItemData[] = [
+      {
+        type: 'ungroup',
+        text: '取消组合',
+      },
+    ]
+    menuListData.value.list = ungroup.concat(menuListData.value.list)
+  }
+  showMenuBg.value = true
+  // document.getElementById('menu-bg').addEventListener('click', this.closeMenu, false)
+  let mx = e.pageX
+  let my = e.pageY
+  let listWidth = 120
+  if (mx + listWidth > window.innerWidth) {
+    mx -= listWidth
+  }
+  let listHeight = (14 + 10) * menuListData.value.list.length + 10
+  if (my + listHeight > window.innerHeight) {
+    my -= listHeight
+  }
+  menuListData.value.left = mx
+  menuListData.value.top = my
+}
+
+function closeMenu() {
+  showMenuBg.value = false
+}
+
+/** 点击菜单触发事件 */
+function selectMenu(type: TWidgetItemData['type']) {
+  switch (type) {
+    case 'copy':
+      store.dispatch('copyWidget')
+      break
+    case 'paste':
+      if (dCopyElement.value.length === 0) {
         return
       }
-      // let target = e.target
-      let target = await getTarget(e.target)
+      store.dispatch('pasteWidget')
+      break
+    case 'index-up':
+      store.dispatch('updateLayerIndex', {
+        uuid: dActiveElement.value.uuid,
+        value: 1,
+        isGroup: dActiveElement.value.isContainer,
+      })
+      break
+    case 'index-down':
+      store.dispatch('updateLayerIndex', {
+        uuid: dActiveElement.value.uuid,
+        value: -1,
+        isGroup: dActiveElement.value.isContainer,
+      })
+      break
+    case 'del':
+      store.dispatch('deleteWidget')
+      break
+    case 'ungroup':
+      store.dispatch('ungroup', dActiveElement.value.uuid)
+      break
+  }
+  closeMenu()
+}
 
-      let type = target.getAttribute('data-type')
-      if (type) {
-        let uuid = target.getAttribute('data-uuid') // 设置选中元素
-
-        if (uuid !== '-1' && !this.dAltDown) {
-          let widget = this.dWidgets.find((item: any) => item.uuid === uuid)
-          if (widget.parent !== '-1' && widget.parent !== this.dActiveElement.uuid && widget.parent !== this.dActiveElement.parent) {
-            uuid = widget.parent
-          }
-        }
-        this.selectWidget({
-          uuid: uuid || '-1',
-        })
-        this.showMenu(e)
-      }
-    },
-    showMenu(e: any) {
-      let isPage = this.dActiveElement.uuid === '-1'
-      this.menuList.list = isPage ? this.pageMenu : this.widgetMenu
-      if (this.dActiveElement.isContainer) {
-        let ungroup = [
-          {
-            type: 'ungroup',
-            text: '取消组合',
-          },
-        ]
-        this.menuList.list = ungroup.concat(this.menuList.list)
-      }
-      this.showMenuBg = true
-      // document.getElementById('menu-bg').addEventListener('click', this.closeMenu, false)
-      let mx = e.pageX
-      let my = e.pageY
-      let listWidth = 120
-      if (mx + listWidth > window.innerWidth) {
-        mx -= listWidth
-      }
-      let listHeight = (14 + 10) * this.menuList.list.length + 10
-      if (my + listHeight > window.innerHeight) {
-        my -= listHeight
-      }
-      this.menuList.left = mx
-      this.menuList.top = my
-    },
-    closeMenu() {
-      this.showMenuBg = false
-      // document.getElementById('menu-bg').removeEventListener('click', this.closeMenu, false)
-    },
-    selectMenu(type) {
-      switch (type) {
-        case 'copy':
-          this.copyWidget()
-          break
-        case 'paste':
-          if (this.dCopyElement.length === 0) {
-            return
-          }
-          this.pasteWidget()
-          break
-        case 'index-up':
-          this.updateLayerIndex({
-            uuid: this.dActiveElement.uuid,
-            value: 1,
-            isGroup: this.dActiveElement.isContainer,
-          })
-          break
-        case 'index-down':
-          this.updateLayerIndex({
-            uuid: this.dActiveElement.uuid,
-            value: -1,
-            isGroup: this.dActiveElement.isContainer,
-          })
-          break
-        case 'del':
-          this.deleteWidget()
-          break
-        case 'ungroup':
-          this.ungroup(this.dActiveElement.uuid)
-          break
-      }
-      this.closeMenu()
-    },
-  },
-})
 </script>
 
 <style lang="less" scoped>
