@@ -5,6 +5,8 @@
  * @LastEditors: ShawnPhang
  * @LastEditTime: 2022-04-08 10:28:47
  */
+import { App } from "vue"
+
 function capitalizeFirstLetter(string: string) {
   return string.charAt(0).toUpperCase() + string.slice(1)
 }
@@ -15,9 +17,10 @@ const exclude = ['settings', 'layout']
 const regex = RegExp('.*^(?!.*?(' + exclude.join('|') + ')).*\\.vue$')
 
 // const requireComponent = require.context('.', true, /\.vue$/) // 找到components文件夹下以.vue命名的文件
-const requireComponent = import.meta.globEager('./**/*.vue')
 
-function guide(Vue: Type.Object) {
+const requireComponent = import.meta.glob('./**/*.vue', { eager: true })
+
+function guide(Vue: App) {
   for (const fileName in requireComponent) {
     if (regex.test(fileName)) {
       const componentConfig = requireComponent[fileName]
