@@ -7,206 +7,247 @@
 -->
 <template>
   <div class="wrap">
-    <el-tabs v-model="tabActiveName" :stretch="true" class="tabs" @tab-change="tabChange">
+    <el-tabs v-model="state.tabActiveName" :stretch="true" class="tabs" @tab-change="tabChange">
       <el-tab-pane label="资源管理" name="pics"> </el-tab-pane>
       <el-tab-pane label="我的作品" name="design"> </el-tab-pane>
     </el-tabs>
-    <div v-show="tabActiveName === 'pics'">
-      <uploader v-model="percent" class="upload" @done="uploadDone">
+    <div v-show="state.tabActiveName === 'pics'">
+      <uploader v-model="state.percent" class="upload" @done="uploadDone">
         <el-button class="upload-btn" plain>上传图片 <i class="iconfont icon-upload" /></el-button>
       </uploader>
       <el-button class="upload-btn upload-psd" plain type="primary" @click="openPSD">上传 PSD 模板</el-button>
       <div style="margin: 1rem; height: 100vh">
-        <photo-list ref="imgListRef" :edit="editOptions.photo" :isDone="isDone" :listData="imgList" @load="load" @drag="dragStart" @select="selectImg" />
+        <photo-list
+          ref="imgListRef" 
+          :edit="state.editOptions.photo" :isDone="state.isDone"
+          :listData="state.imgList"
+          @load="load" @drag="dragStart" @select="selectImg"
+        />
       </div>
     </div>
-    <div v-show="tabActiveName === 'design'" class="wrap">
+    <div v-show="state.tabActiveName === 'design'" class="wrap">
       <ul ref="listRef" v-infinite-scroll="loadDesign" class="infinite-list" :infinite-scroll-distance="150" style="overflow: auto">
-        <img-water-fall :edit="editOptions.works" :listData="designList" @select="selectDesign" />
+        <img-water-fall :edit="state.editOptions.works" :listData="state.designList" @select="selectDesign" />
         <!-- <div v-show="loading" class="loading"><i class="el-icon-loading"></i>拼命加载中..</div> -->
-        <div v-show="isDone" class="loading">全部加载完毕</div>
+        <div v-show="state.isDone" class="loading">全部加载完毕</div>
       </ul>
     </div>
   </div>
 </template>
 
-<script lang="ts">
-import { defineComponent, reactive, toRefs, watch, nextTick, ref, onMounted } from 'vue'
-import { ElTabPane, ElTabs } from 'element-plus'
+<script lang="ts" setup>
+import { reactive, toRefs, watch, nextTick, ref, onMounted, defineProps, defineExpose } from 'vue'
+import { ElTabPane, ElTabs, TabPaneName } from 'element-plus'
 import { useRouter } from 'vue-router'
 import { useStore } from 'vuex'
 import uploader from '@/components/common/Uploader'
 import api from '@/api'
 import wImage from '../../widgets/wImage/wImage.vue'
-import setImageData from '@/common/methods/DesignFeatures/setImage'
+import setImageData, { TItem2DataParam } from '@/common/methods/DesignFeatures/setImage'
 import useConfirm from '@/common/methods/confirm'
+import { TGetImageListResult } from '@/api/material'
+import photoList from './components/photoList.vue'
+import imgWaterFall from './components/imgWaterFall.vue'
+import { TUploadDoneData } from '@/components/common/Uploader/index.vue'
+import { IGetTempListData } from '@/api/home'
 
-export default defineComponent({
-  components: { uploader, ElTabPane, ElTabs },
-  props: ['active'],
-  setup(props) {
-    const router = useRouter()
-    const store = useStore()
-    const state: any = reactive({
-      prePath: 'user',
-      percent: { num: 0 }, // 当前上传进度
-      imgList: [],
-      designList: [],
-      isDone: false,
-      editOptions: [],
-      listRef: null,
-      imgListRef: null,
-      tabActiveName: '',
-    })
-    let loading = false
-    let page = 0
-    let listPage = 0
+type TProps = {
+  active?: number
+}
 
-    const load = (init: boolean) => {
-      if (init) {
-        state.imgList = []
-        page = 0
-        state.isDone = false
-      }
-      if (state.isDone || loading) {
-        return
-      }
-      loading = true
-      page += 1
-      api.material.getMyPhoto({ page }).then(({ list }: any) => {
-        list.length <= 0 ? (state.isDone = true) : (state.imgList = state.imgList.concat(list))
-        setTimeout(() => {
-          loading = false
-          checkHeight(state.imgListRef.getRef(), load)
-        }, 100)
-      })
-    }
-    const loadDesign = (init: boolean = false) => {
-      if (init) {
-        state.designList = []
-        listPage = 0
-        state.isDone = false
-      }
-      if (state.isDone || loading) {
-        return
-      }
-      loading = true
-      listPage += 1
-      api.home.getMyDesign({ page: listPage, pageSize: 10 }).then(({ list }: any) => {
-        list.length <= 0
-          ? (state.isDone = true)
-          : (state.designList = state.designList.concat(
-              list.map((x: any) => {
-                x.cover = x.cover + '?r=' + Math.random()
-                return x
-              }),
-            ))
-        setTimeout(() => {
-          loading = false
-          checkHeight(state.listRef, loadDesign)
-        }, 100)
-      })
-    }
+type TState = {
+  prePath: string,
+  percent: { num: number }, // 当前上传进度
+  imgList: IGetTempListData[],
+  designList: IGetTempListData[],
+  isDone: boolean,
+  editOptions: Record<string, any>,
+  tabActiveName: string,
+}
 
-    function checkHeight(el: any, loadFn: Function) {
-      // 检查高度是否占满，否则继续请求下一页
-      const isLess = el.offsetHeight > el.firstElementChild.offsetHeight
-      isLess && loadFn()
-    }
+const props = defineProps<TProps>()
 
-    onMounted(() => {
-      load(true)
-      nextTick(() => {
-        state.tabActiveName = 'pics'
-      })
-    })
+const router = useRouter()
+const store = useStore()
+const listRef = ref<HTMLElement | null>(null)
+const imgListRef = ref<typeof photoList | null>(null)
 
-    const selectImg = async (index: number) => {
-      const item: any = state.imgList[index]
-      store.commit('setShowMoveable', false) // 清理掉上一次的选择
-      let setting = JSON.parse(JSON.stringify(wImage.setting))
-      const img: any = await setImageData(item)
-      setting.width = img.width
-      setting.height = img.height // parseInt(100 / item.value.ratio, 10)
-      setting.imgUrl = item.url
-      const { width: pW, height: pH } = store.getters.dPage
-      setting.left = pW / 2 - img.width / 2
-      setting.top = pH / 2 - img.height / 2
-      store.dispatch('addWidget', setting)
-    }
-    const deleteImg = async ({ i, item }: any) => {
-      store.commit('setShowMoveable', false) // 清理掉上一次的选择框
-      const isPass = await useConfirm('警告', '删除后不可找回，已引用资源将会失效，请谨慎操作', 'warning')
-      if (!isPass) {
-        return false
-      }
-      const arr = item.url.split('/')
-      let key = arr.splice(3, arr.length - 1).join('/')
-      api.material.deleteMyPhoto({ id: item.id, key })
-      state.imgListRef.delItem(i) // 通知标记
-    }
-    const deleteWorks = async ({ i, item }: any) => {
-      const isPass = await useConfirm('警告', '删除后不可找回，请确认操作', 'warning')
-      if (isPass) {
-        await api.material.deleteMyWorks({ id: item.id })
-        setTimeout(() => {
-          router.push({ path: '/home', query: {  }, replace: true })
-          loadDesign(true)
-        }, 300);
-      }
-    }
-    state.editOptions = {
-      photo: [
-      {
-        name: '删除',
-        fn: deleteImg,
-      },
-    ],works: [
-      {
-        name: '删除',
-        fn: deleteWorks,
-      },
-    ]
-    }
-    const dragStart = (index: number) => {
-      const item = state.imgList[index]
-      store.commit('selectItem', { data: { value: item }, type: 'image' })
-    }
-    const uploadDone = async (res: any) => {
-      await api.material.addMyPhoto(res)
-      state.imgList = []
-      load(true)
-    }
+const state = reactive<TState>({
+  prePath: 'user',
+  percent: { num: 0 }, // 当前上传进度
+  imgList: [],
+  designList: [],
+  isDone: false,
+  editOptions: [],
+  tabActiveName: '',
+})
 
-    const tabChange = (tabName: string) => {
-      if (tabName === 'design') {
-        loadDesign(true)
-      }
-    }
+let loading = false
+let page = 0
+let listPage = 0
 
-    const selectDesign = async (item: any) => {
-      // const { id }: any = state.designList[index]
-      const { id }: any = item
-      window.open(`${window.location.protocol + '//' + window.location.host}/home?id=${id}`)
+const load = (init?: boolean) => {
+  if (init) {
+    state.imgList = []
+    page = 0
+    state.isDone = false
+  }
+  if (state.isDone || loading) {
+    return
+  }
+  loading = true
+  page += 1
+  api.material.getMyPhoto({ page }).then(({ list }) => {
+    
+    if (list.length <= 0) {
+      state.isDone = true
+    } else {
+      state.imgList = state.imgList.concat(list)
     }
+    setTimeout(() => {
+      loading = false
+      if (!imgListRef.value) return
+      checkHeight(imgListRef.value.getRef(), load)
+    }, 100)
+  })
+}
 
-    const openPSD = () => {
-      window.open(router.resolve('/psd').href, '_blank')
-    }
+const loadDesign = (init: boolean = false) => {
+  if (init) {
+    state.designList = []
+    listPage = 0
+    state.isDone = false
+  }
+  if (state.isDone || loading) {
+    return
+  }
+  loading = true
+  listPage += 1
+  api.home.getMyDesign({ page: listPage, pageSize: 10 }).then(({ list }) => {
+    list.length <= 0
+      ? (state.isDone = true)
+      : (state.designList = state.designList.concat(
+          list.map((x) => {
+            x.cover = x.cover + '?r=' + Math.random()
+            return x
+          }),
+        ))
+    setTimeout(() => {
+      loading = false
+      if (!listRef.value) return
+      checkHeight(listRef.value, loadDesign)
+    }, 100)
+  })
+}
 
-    return {
-      ...toRefs(state),
-      selectDesign,
-      loadDesign,
-      load,
-      uploadDone,
-      selectImg,
-      deleteImg,
-      dragStart,
-      tabChange,
-      openPSD,
-    }
-  },
+function checkHeight(el: HTMLElement, loadFn: Function) {
+  // 检查高度是否占满，否则继续请求下一页
+  if (el.offsetHeight && el.firstElementChild) {
+    const isLess = el.offsetHeight > (el.firstElementChild as HTMLElement).offsetHeight
+    isLess && loadFn()
+  }
+}
+
+onMounted(() => {
+  load(true)
+  nextTick(() => {
+    state.tabActiveName = 'pics'
+  })
+})
+
+const selectImg = async (index: number) => {
+  const item = state.imgList[index]
+  store.commit('setShowMoveable', false) // 清理掉上一次的选择
+  let setting = JSON.parse(JSON.stringify(wImage.setting))
+  const img = await setImageData(item)
+  setting.width = img.width
+  setting.height = img.height // parseInt(100 / item.value.ratio, 10)
+  setting.imgUrl = item.url
+  const { width: pW, height: pH } = store.getters.dPage
+  setting.left = pW / 2 - img.width / 2
+  setting.top = pH / 2 - img.height / 2
+  store.dispatch('addWidget', setting)
+}
+
+type controlImgParam = {
+  i: number
+  item: Required<TItem2DataParam>
+}
+
+const deleteImg = async ({ i, item }: controlImgParam) => {
+  store.commit('setShowMoveable', false) // 清理掉上一次的选择框
+  const isPass = await useConfirm('警告', '删除后不可找回，已引用资源将会失效，请谨慎操作', 'warning')
+  if (!isPass) {
+    return false
+  }
+  const arr = item.url.split('/')
+  let key = arr.splice(3, arr.length - 1).join('/')
+  api.material.deleteMyPhoto({ id: item.id, key })
+  if (!imgListRef.value) return
+  imgListRef.value.delItem(i) // 通知标记
+}
+const deleteWorks = async ({ i, item }: controlImgParam) => {
+  const isPass = await useConfirm('警告', '删除后不可找回，请确认操作', 'warning')
+  if (isPass) {
+    await api.material.deleteMyWorks({ id: item.id })
+    setTimeout(() => {
+      router.push({ path: '/home', query: {  }, replace: true })
+      loadDesign(true)
+    }, 300);
+  }
+}
+
+state.editOptions = {
+  photo: [
+    {
+      name: '删除',
+      fn: deleteImg,
+    },
+  ],
+  works: [
+    {
+      name: '删除',
+      fn: deleteWorks,
+    },
+  ]
+}
+
+const dragStart = (index: number) => {
+  const item = state.imgList[index]
+  store.commit('selectItem', { data: { value: item }, type: 'image' })
+}
+const uploadDone = async (res: TUploadDoneData) => {
+  await api.material.addMyPhoto(res)
+  state.imgList = []
+  load(true)
+}
+
+const tabChange = (tabName: TabPaneName) => {
+  if (tabName === 'design') {
+    loadDesign(true)
+  }
+}
+
+const selectDesign = async (item: IGetTempListData) => {
+  // const { id }: any = state.designList[index]
+  const { id } = item
+  window.open(`${window.location.protocol + '//' + window.location.host}/home?id=${id}`)
+}
+
+const openPSD = () => {
+  window.open(router.resolve('/psd').href, '_blank')
+}
+
+defineExpose({
+  selectDesign,
+  loadDesign,
+  load,
+  uploadDone,
+  selectImg,
+  deleteImg,
+  dragStart,
+  tabChange,
+  openPSD,
 })
 </script>
 
