@@ -2,8 +2,8 @@
  * @Author: ShawnPhang
  * @Date: 2023-10-04 02:04:04
  * @Description: 列表分类头部
- * @LastEditors: ShawnPhang <https://m.palxp.cn>
- * @LastEditTime: 2023-10-04 02:30:59
+ * @LastEditors: ShawnPhang <https://m.palxp.cn>, Jeremy Yu <https://github.com/JeremyYu-cn>
+ * @Date: 2024-03-06 21:16:00
 -->
 <template>
   <div v-if="!isBack" class="content__wrap">
@@ -21,22 +21,34 @@
   </span>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue'
+<script lang="ts" setup>
+import { defineProps, defineEmits } from 'vue'
 
-export default defineComponent({
-  props: ['types', 'isBack'],
-  emits: ['select', 'back'],
-  setup(props, { emit }) {
-    const select = (item: any) => {
-      emit('select', item)
-    }
-    const back = () => {
-      emit('back')
-    }
-    return { select, back }
-  },
-})
+export type TClassHeaderTypeData = {
+  name: string
+}
+
+type TProps = {
+  types: TClassHeaderTypeData[]
+  isBack?: boolean
+}
+
+type TEmits = {
+  (event: 'select', data: string[]): void
+  (event: 'back'): void
+}
+
+const { types, isBack } = defineProps<TProps>()
+const emit = defineEmits<TEmits>()
+
+const select = (item: any) => {
+  emit('select', item)
+}
+const back = () => {
+  emit('back')
+}
+
+defineExpose({ select, back })
 </script>
 
 <style lang="less" scoped>
