@@ -2,8 +2,8 @@
  * @Author: ShawnPhang
  * @Date: 2023-10-04 19:12:40
  * @Description: 图片描述ToolTip
- * @LastEditors: ShawnPhang <https://m.palxp.cn>
- * @LastEditTime: 2023-10-04 22:51:06
+ * @LastEditors: ShawnPhang <https://m.palxp.cn>, Jeremy Yu <https://github.com/JeremyYu-cn>
+ * @Date: 2024-03-06 21:16:00
 -->
 <template>
   <el-tooltip :disabled="!detail.author" :offset="1" effect="light" placement="bottom-start" :hide-after="0" :enterable="false" raw-content>
@@ -17,15 +17,18 @@
   </el-tooltip>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue'
+<script lang="ts" setup>
+import { defineProps } from 'vue'
 
-export default defineComponent({
-  props: {
-    detail: {},
-  },
-  setup() {
-    return {}
-  },
-})
+export type TImageTipDetailData = {
+  author: string
+  description: string
+}
+
+type Tprops = {
+  detail: TImageTipDetailData
+}
+
+const { detail } = defineProps<Tprops>()
+
 </script>
