@@ -2,8 +2,8 @@
  * @Author: ShawnPhang
  * @Date: 2022-01-27 11:05:48
  * @Description:  
- * @LastEditors: ShawnPhang <https://m.palxp.cn>
- * @LastEditTime: 2023-10-04 01:53:10
+ * @LastEditors: ShawnPhang <https://m.palxp.cn>, Jeremy Yu <https://github.com/JeremyYu-cn>
+ * @Date: 2024-03-06 21:16:00
 -->
 <template>
   <div class="search__wrap">
@@ -11,66 +11,85 @@
       <div class="search__type"><i class="iconfont icon-ego-caidan" /></div>
       <template #dropdown>
         <el-dropdown-menu>
-          <el-dropdown-item v-for="type in materialCates" :key="type.id" @click="action('change', type, type.id)">
-            <span :class="['cate__text', { 'cate--select': +currentIndex === type.id }]">{{ type.name }}</span>
+          <el-dropdown-item
+            v-for="type in state.materialCates" :key="type.id"
+            @click="action('change', type, type.id)"
+          >
+            <span :class="['cate__text', { 'cate--select': + state.currentIndex === type.id }]">{{ type.name }}</span>
           </el-dropdown-item>
         </el-dropdown-menu>
       </template>
     </el-dropdown>
     <span v-else style="width: 1rem"></span>
 
-    <el-input v-model="searchValue" size="large" placeholder="输入关键词搜索" class="input-with-select">
+    <el-input v-model="state.searchValue" size="large" placeholder="输入关键词搜索" class="input-with-select">
       <template #append>
         <el-button><i class="iconfont icon-search"></i></el-button>
       </template>
     </el-input>
   </div>
 </template>
-<script lang="ts">
-import { defineComponent, reactive, toRefs, watch } from 'vue'
+<script lang="ts" setup>
+import { reactive, toRefs, watch, defineProps, defineEmits, defineExpose } from 'vue'
 import { ElDropdown, ElDropdownItem, ElDropdownMenu } from 'element-plus'
 import { useRoute } from 'vue-router'
 import api from '@/api'
 
-export default defineComponent({
-  components: { ElDropdown, ElDropdownItem, ElDropdownMenu },
-  props: ['type', 'modelValue'],
-  emits: ['update:modelValue'],
-  setup(props, context) {
-    const route = useRoute()
-    const state: any = reactive({
-      searchValue: '',
-      materialCates: [],
-      currentIndex: 1,
-    })
+type TProps = {
+  type?: string
+  modelValue?: string
+}
 
-    if (props.type != 'none') {
-      api.home.getCategories({ type: 1 }).then((list: any) => {
-        list.unshift({ id: 0, name: '全部' })
-        state.materialCates = list
-        const { cate } = route.query
-        cate && (state.currentIndex = cate)
-        cate && action('change', state.materialCates[Number(cate)], Number(cate))
-      })
-    }
+type TEmits = {
+  (event: 'update:modelValue', data: string): void
+  (event: 'change', data: TMaterialCatesData): void
+}
 
-    watch(
-      () => state.searchValue,
-      () => {
-        context.emit('update:modelValue', state.searchValue)
-      },
-    )
+type TMaterialCatesData = {id: string | number, name: string}
 
-    function action(fn: string, type: any, currentIndex: number | string) {
-      currentIndex && (state.currentIndex = currentIndex)
-      context.emit(fn, type)
-    }
-    return {
-      ...toRefs(state),
-      action,
-    }
-  },
+type TState = {
+  searchValue: string
+  materialCates: TMaterialCatesData[]
+  currentIndex: number | string
+}
+
+const props = defineProps<TProps>()
+
+const emit = defineEmits<TEmits>()
+
+const route = useRoute()
+const state = reactive<TState>({
+  searchValue: '',
+  materialCates: [],
+  currentIndex: 1,
 })
+
+if (props.type != 'none') {
+  api.home.getCategories({ type: 1 }).then((list: any) => {
+    list.unshift({ id: 0, name: '全部' })
+    state.materialCates = list
+    const { cate } = route.query
+    cate && (state.currentIndex = cate as string)
+    cate && action('change', state.materialCates[Number(cate)], Number(cate))
+  })
+}
+
+watch(
+  () => state.searchValue,
+  () => {
+    emit('update:modelValue', state.searchValue)
+  },
+)
+
+function action(fn: 'change', type: TMaterialCatesData, currentIndex: number | string) {
+  currentIndex && (state.currentIndex = currentIndex)
+  emit(fn, type)
+}
+
+defineExpose({
+  action
+})
+
 </script>
 
 <style lang="less" scoped>
