@@ -2,8 +2,8 @@
  * @Author: ShawnPhang
  * @Date: 2022-03-03 14:13:16
  * @Description:
- * @LastEditors: ShawnPhang <site: book.palxp.com>
- * @LastEditTime: 2023-06-29 15:11:46
+ * @LastEditors: ShawnPhang <https://m.palxp.cn>, Jeremy Yu <https://github.com/JeremyYu-cn>
+ * @LastEditTime: 2024-02-26 17:54:00
  */
 import { createApp } from 'vue'
 import App from './App.vue'
@@ -20,7 +20,7 @@ elementConfig.components.forEach((component) => {
   app.component(component.name, component)
 })
 
-elementConfig.plugins.forEach((plugin: any) => {
+elementConfig.plugins.forEach((plugin) => {
   app.use(plugin)
 })
 
