@@ -1,4 +1,4 @@
-/// <reference lib="dom" />
+// / <reference lib="dom" />
 
 interface ImportMeta {
     url: string
@@ -23,13 +23,14 @@ interface ImportMeta {
       on(event: string, cb: (...args: any[]) => void): void
     }
   
-    readonly env: ImportMetaEnv
+    // readonly env: ImportMetaEnv
+    
   
-    glob(pattern: string): Record<
+    glob(pattern: string, { eager: boolean }): Record<
       string,
-      () => Promise<{
+      {
         [key: string]: any
-      }>
+      }
     >
   
     globEager(pattern: string): Record<
