@@ -6,9 +6,10 @@
  * @LastEditTime: 2023-09-19 17:32:40
  */
 export default class PointImg {
-  private canvas: any
-  private cvs: any
-  constructor(img: any) {
+  private canvas: HTMLCanvasElement | undefined
+  private cvs: CanvasRenderingContext2D | null | undefined
+
+  constructor(img: HTMLImageElement) {
     if (img.src) {
       try {
         this.canvas = document.createElement('canvas')
@@ -16,6 +17,8 @@ export default class PointImg {
         this.canvas.height = img.height
         img.crossOrigin = 'Anonymous'
         this.cvs = this.canvas.getContext('2d')
+        if (!this.cvs) return
+
         this.cvs.drawImage(img, 0, 0, img.width, img.height)
       } catch (error) {
         console.log(error)
@@ -28,25 +31,27 @@ export default class PointImg {
      * @param y Number y坐标起点
      * @return color Object 包含颜色的rgba #16进制颜色
      */
-    const color: any = {}
+    const color: Record<string, string> = {}
     try {
-      const obj = this.cvs.getImageData(x, y, 1, 1)
-      const arr = obj.data.toString().split(',')
+      if (this.cvs) {
+        const obj = this.cvs.getImageData(x, y, 1, 1)
+        const arr = obj.data.toString().split(',')
 
-      let first = parseInt(arr[0], 10).toString(16)
-      first = first.length === 2 ? first : first + first
+        let first = parseInt(arr[0], 10).toString(16)
+        first = first.length === 2 ? first : first + first
 
-      let second = parseInt(arr[1], 10).toString(16)
-      second = second.length === 2 ? second : second + second
+        let second = parseInt(arr[1], 10).toString(16)
+        second = second.length === 2 ? second : second + second
 
-      let third = parseInt(arr[2], 10).toString(16)
-      third = third.length === 2 ? third : third + third
+        let third = parseInt(arr[2], 10).toString(16)
+        third = third.length === 2 ? third : third + third
 
-      let last = parseInt(arr.pop(), 10) / 255
-      last = Number(last.toFixed(0))
+        let last = parseInt(arr.pop() || '0', 10) / 255
+        last = Number(last.toFixed(0))
 
-      color['rgba'] = 'rgba(' + arr.join(',') + ',' + last + ')'
-      color['#'] = '#' + first + second + third
+        color['rgba'] = 'rgba(' + arr.join(',') + ',' + last + ')'
+        color['#'] = '#' + first + second + third
+      }
     } catch (error) {
       // console.log('此为解析图片点位异常')
     }
