@@ -2,8 +2,8 @@
  * @Author: ShawnPhang
  * @Date: 2023-08-23 17:37:16
  * @Description: 提取字体子集
- * @LastEditors: ShawnPhang <https://m.palxp.cn>
- * @LastEditTime: 2023-10-14 18:31:29
+ * @LastEditors: ShawnPhang <https://m.palxp.cn>, Jeremy Yu <https://github.com/JeremyYu-cn>
+ * @LastEditTime: 2024-02-27 10:32:00
  */
 /**
  * 只有ttf/otf这种原始字体支持提取，如果服务端不支持该功能请设置false，以保证页面能加载字体。
