@@ -13,19 +13,10 @@
   </tool-tip>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue'
+<script lang="ts" setup>
 import toolTip from '@/components/common/PopoverTip.vue'
 
-export default defineComponent({
-  components: { toolTip },
-  setup() {
-    const content = '本站为个人项目，所使用素材图片等均为网络收集而来，下载之作品仅供学习研究或欣赏目的而使用，无法提供商用授权哦。'
-    return {
-      content,
-    }
-  },
-})
+const content = '本站为个人项目，所使用素材图片等均为网络收集而来，下载之作品仅供学习研究或欣赏目的而使用，无法提供商用授权哦。'
 </script>
 
 <style lang="less" scoped>
